// Server/src/user/entities/user.interface.ts
import { Document } from 'mongoose';

export interface User extends Document {
    name: string

    lastName: string
  
    email: string
  
    password: string
  
    profileImage: string
  
    age: number
  
    role: string
  
    score: {
      point: number
      date: Date
    }

    deleted_at?: Date
}